Use user argument for passdb dict lookup
Based on the changes from dovecot it's required to handle the dict
arguments correctly. The provided arguments contains the key and
username seperated by tab. The key seperation is specified via
`password_key` and is `passdb/%u` in core-mbox.

The example code (in perl) contains a split into key and user. We should
use the user for the authentication against redis so the value of
`password_key` could be changed to `passdb` without any variable
information in the key.

References:
 https://doc.dovecot.org/configuration_manual/authentication/dict/